Refactor SLH-DSA tests
