+++ src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/SlhDsa/SlhDsaAlgorithmTests.cs
@@ -5,7 +5,7 @@ using Xunit;
 
 namespace System.Security.Cryptography.SLHDsa.Tests
 {
-    public class SlhDsaAlgorithmTests
+    public sealed class SlhDsaAlgorithmTests
     {
         [Fact]
         public static void AlgorithmsHaveExpectedParameters()
