+++ src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/SlhDsa/SlhDsaTestHelpers.cs
@@ -1,0 +1,62 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+using System.Drawing;
+using System.Formats.Asn1;
+using System.Linq;
+using System.Reflection.Emit;
+using System.Security.Cryptography.Asn1;
+using System.Text;
+using Xunit;
+using Xunit.Sdk;
+
+namespace System.Security.Cryptography.SLHDsa.Tests
+{
+    internal static class SlhDsaTestHelpers
+    {
+        /// <summary>
+        /// Gets the negation of <see cref="SlhDsa.IsSupported"/>. This can be used to conditionally skip tests.
+        /// </summary>
+        public static bool IsNotSupported => !SlhDsa.IsSupported;
+
+        public static void VerifyDisposed(SlhDsa slhDsa)
+        {
+            // A signature-sized buffer can be reused for keys as well
+            byte[] tempBuffer = new byte[slhDsa.Algorithm.SignatureSizeInBytes];
+            PbeParameters pbeParameters = new PbeParameters(PbeEncryptionAlgorithm.Aes256Cbc, HashAlgorithmName.SHA256, 32);
+
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.SignData([], tempBuffer, []));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.VerifyData([], tempBuffer, []));
+
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportEncryptedPkcs8PrivateKey(ReadOnlySpan<byte>.Empty, pbeParameters));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportEncryptedPkcs8PrivateKey(ReadOnlySpan<char>.Empty, pbeParameters));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportEncryptedPkcs8PrivateKeyPem(ReadOnlySpan<byte>.Empty, pbeParameters));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportEncryptedPkcs8PrivateKeyPem(ReadOnlySpan<char>.Empty, pbeParameters));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportPkcs8PrivateKey());
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportPkcs8PrivateKeyPem());
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportSlhDsaPublicKey(tempBuffer));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportSlhDsaSecretKey(tempBuffer));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportSubjectPublicKeyInfo());
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.ExportSubjectPublicKeyInfoPem());
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.TryExportEncryptedPkcs8PrivateKey(ReadOnlySpan<byte>.Empty, pbeParameters, [], out _));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.TryExportEncryptedPkcs8PrivateKey(ReadOnlySpan<char>.Empty, pbeParameters, [], out _));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.TryExportPkcs8PrivateKey(tempBuffer, out _));
+            Assert.Throws<ObjectDisposedException>(() => slhDsa.TryExportSubjectPublicKeyInfo([], out _));
+        }
+
+        public static void AssertSameBuffer(ReadOnlySpan<byte> expected, ReadOnlySpan<byte> actual)
+        {
+            if (expected.Length != actual.Length)
+            {
+                Assert.Fail("Expected buffers to have same length. " +
+                    $"The first buffer's length {expected.Length} does not match the second buffer's length {actual.Length}.");
+            }
+
+            if (!expected.Overlaps(actual, out int offset) || offset != 0)
+            {
+                Assert.Fail("Expected buffers to be the same memory location, but were not.");
+            }
+        }
+    }
+}
