+++ src/libraries/System.Security.Cryptography/tests/SlhDsaOpenSslConstructionTests.NotSupported.cs
@@ -5,31 +5,13 @@ using Xunit;
 
 namespace System.Security.Cryptography.SLHDsa.Tests
 {
-    public class SlhDsaOpenSslConstructionTests : SlhDsaConstructionTestsBase
+    public static class SlhDsaOpenSslConstructionTests
     {
         [Fact]
         public static void SlhDsaOpenSsl_NotSupportedOnNonUnixPlatforms()
         {
             Assert.Throws<PlatformNotSupportedException>(() => new SlhDsaOpenSsl(null));
             Assert.Throws<PlatformNotSupportedException>(() => new SlhDsaOpenSsl(new SafeEvpPKeyHandle()));
-        }
-
-        protected override SlhDsa GenerateKey(SlhDsaAlgorithm algorithm)
-        {
-            Assert.Fail();
-            throw new PlatformNotSupportedException();
-        }
-
-        protected override SlhDsa ImportSlhDsaPublicKey(SlhDsaAlgorithm algorithm, ReadOnlySpan<byte> source)
-        {
-            Assert.Fail();
-            throw new PlatformNotSupportedException();
-        }
-
-        protected override SlhDsa ImportSlhDsaSecretKey(SlhDsaAlgorithm algorithm, ReadOnlySpan<byte> source)
-        {
-            Assert.Fail();
-            throw new PlatformNotSupportedException();
         }
     }
 }
